Fixed bug in metal::scatter()

+++ vec.cuh
@@ -50,9 +50,9 @@ __host__ __device__ inline vec3 reflect(const vec3& v, const vec3& n) {
 }
 
 __host__ __device__ inline vec3 refract(const vec3& uv, const vec3& n, const float etai_over_etat) {
-    const auto cos_theta = std::fmin(dot(-uv, n), 1.0f);
+    const auto cos_theta = fmin(dot(-uv, n), 1.0f);
     const vec3 r_out_perp =  etai_over_etat * (uv + cos_theta*n);
-    const vec3 r_out_parallel = -std::sqrt(std::fabs(1.0f - glm::dot(r_out_perp,r_out_perp))) * n;
+    const vec3 r_out_parallel = -sqrtf(fabs(1.0f - glm::dot(r_out_perp,r_out_perp))) * n;
     return r_out_perp + r_out_parallel;
 }
